// App/src/app/services/didac/didac-client.generated.ts
export class EducationCard implements IEducationCard {
    deckId?: string | undefined;
    cardId?: string | undefined;
    playerCardId?: string | undefined;
    sessionId?: string | undefined;
    sessionCardId?: string | undefined;
    content?: string | undefined;

    constructor(data?: IEducationCard) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.deckId = _data["deckId"];
            this.cardId = _data["cardId"];
            this.playerCardId = _data["playerCardId"];
            this.sessionId = _data["sessionId"];
            this.sessionCardId = _data["sessionCardId"];
            this.content = _data["content"];
        }
    }

    static fromJS(data: any): EducationCard {
        data = typeof data === 'object' ? data : {};
        let result = new EducationCard();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["deckId"] = this.deckId;
        data["cardId"] = this.cardId;
        data["playerCardId"] = this.playerCardId;
        data["sessionId"] = this.sessionId;
        data["sessionCardId"] = this.sessionCardId;
        data["content"] = this.content;
        return data; 
    }
}

export interface IEducationCard {
    deckId?: string | undefined;
    cardId?: string | undefined;
    playerCardId?: string | undefined;
    sessionId?: string | undefined;
    sessionCardId?: string | undefined;
    content?: string | undefined;
}